+++ src/column_statistics.rs
@@ -1,3 +1,4 @@
+mod delete;
 mod load;
 mod round;
 mod save;